Fix otp text color
