+++ app/src/main/java/org/pragyan/dalal18/adapter/pagerAdapters/HelpPagerAdapter.kt
@@ -1,17 +1,16 @@
-package org.pragyan.dalal18.adapter
+package org.pragyan.dalal18.adapter.pagerAdapters
 
-import org.pragyan.dalal18.fragment.mortgage.MortgageFragment
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
-import org.pragyan.dalal18.fragment.mortgage.RetrieveFragment
+import org.pragyan.dalal18.fragment.help.FaqFragment
+import org.pragyan.dalal18.fragment.help.GettingStartedFragment
 
-class MortgagePagerAdapter(fm: FragmentManager) : FragmentPagerAdapter(fm) {
-
+class HelpPagerAdapter(fm : FragmentManager) : FragmentPagerAdapter(fm){
     override fun getItem(position: Int): Fragment {
         return when (position) {
-            0 -> MortgageFragment()
-            else -> RetrieveFragment()
+            0 -> GettingStartedFragment()
+            else -> FaqFragment()
         }
     }
 
@@ -19,8 +18,8 @@ class MortgagePagerAdapter(fm: FragmentManager) : FragmentPagerAdapter(fm) {
 
     override fun getPageTitle(position: Int): CharSequence? {
         return when (position) {
-            0 -> "Mortgage"
-            else -> "Retrieve"
+            0 -> "Getting Started"
+            else -> "FAQ"
         }
     }
