+++ app/src/main/java/org/pragyan/dalal18/fragment/help/HelpMainFragment.kt
@@ -10,7 +10,7 @@ import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import kotlinx.android.synthetic.main.fragment_main_help.*
 import org.pragyan.dalal18.R
-import org.pragyan.dalal18.adapter.HelpPagerAdapter
+import org.pragyan.dalal18.adapter.pagerAdapters.HelpPagerAdapter
 
 class HelpMainFragment : Fragment(){
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
