+++ app/src/main/java/org/pragyan/dalal18/fragment/smsVerification/AddPhoneFragment.kt
@@ -4,7 +4,6 @@ import android.app.AlertDialog
 import android.content.Context
 import android.content.SharedPreferences
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
