+++ app/src/main/res/values/strings.xml
@@ -653,6 +653,7 @@
     <string name="india_country_code">91</string>
     <string name="sending_otp">Sending OTP…</string>
     <string name="verifying_otp">Verifying OTP</string>
+    <string name="send_otp">Send OTP</string>
     <string-array name="intervalType">
         <item>1 min</item>
         <item>5 mins</item>
